refactor: implement selectors at finder level
